auth-service/app/requirements/common.txt
alembic==1.8.1
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==22.1.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
flake8==4.0.0
flasgger==0.9.5
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
googleapis-common-protos==1.56.2
greenlet==1.1.2
grpcio==1.48.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.15.0
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.6.1
mistune==2.0.4
opentelemetry-api==1.12.0
opentelemetry-exporter-jaeger==1.12.0
opentelemetry-exporter-jaeger-proto-grpc==1.12.0
opentelemetry-exporter-jaeger-thrift==1.12.0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-flask==0.33b0
opentelemetry-instrumentation-wsgi==0.33b0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-util-http==0.33b0
packaging==21.3
pluggy==0.13.1
protobuf==3.20.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==6.2.5
PyYAML==6.0
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.40
thrift==0.16.0
toml==0.10.2
typing_extensions==4.3.0
Werkzeug==2.2.2
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
httpx
isort==5.10.1